% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{bootstrap}
\alias{bootstrap}
\alias{li}
\alias{ul}
\alias{fa_icon}
\alias{button}
\title{Additional Bootstrap elements}
\usage{
li(..., .noWS = NULL)

ul(..., .noWS = NULL)

fa_icon(..., .noWS = NULL)

button(..., button_type = "default", button_size = "small", .noWS = NULL)
}
\arguments{
\item{...}{Attributes and children of the element. Named arguments become
attributes, and positional arguments become children. Valid children are
tags, single-character character vectors (which become text nodes), raw
HTML (see \code{\link[htmltools]{HTML}}), and \code{html_dependency} objects. You can
also pass lists that contain tags, text nodes, or HTML. To use boolean
attributes, use a named argument with a \code{NA} value. (see example)}

\item{.noWS}{A character vector used to omit some of the whitespace that
would normally be written around this tag. Valid options include
\code{before}, \code{after}, \code{outside}, \code{after-begin},
\code{before-end}, and \code{inside}. Any number of these options can be
specified.}

\item{button_type}{character, one of the standard Bootstrap types}

\item{button_size}{character, size of the button}
}
\description{
Additional Bootstrap elements
}
\examples{
fa_icon("tv")
button("help!", button_type = "primary")
}
\seealso{
\code{\link[htmltools]{tags}}, \code{\link[htmltools]{tag}}
}
