% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{dbConnect_scidb}
\alias{dbConnect_scidb}
\title{Connections to scidb}
\usage{
dbConnect_scidb(dbname, ...)
}
\arguments{
\item{dbname}{the name of the database to which you want to connect}

\item{...}{arguments passed to \code{\link[DBI]{dbConnect}}}
}
\value{
A \code{\link[RMySQL]{MySQLConnection-class}} object
}
\description{
Connect to the \code{scidb} server at Smith College.
}
\details{
This is a public, read-only account. Any abuse will be considered a 
hostile act.
}
\examples{
db_air <- dbConnect_scidb("airlines")
if (require(RMySQL)) {
  summary(db_air)
  dbListTables(db_air)
  if (require(dplyr)) {
    tbl(db_air, "airports")
  }
  dbDisconnect(db_air)
}
}
\seealso{
\code{\link[RMySQL]{MySQLConnection-class}}
}
