% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{standard_button}
\alias{standard_button}
\title{Standard Buttons}
\usage{
standard_button(x, ...)
}
\arguments{
\item{x}{text of the standard}

\item{...}{arguments passed to \code{\link{button}}}
}
\description{
Standard Buttons
}
\examples{
standard_button("aesthetics")
standard_button("markdown")
}
