% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standards}
\alias{standards}
\title{Standards for SDS 192}
\format{
A data frame
\describe{
  \item{standard}{The standard to be met}
  \item{module_id}{The module in which the standard is introduced}
  }
}
\usage{
standards
}
\description{
Standards for SDS 192
}
\keyword{datasets}
